Combine spec files into single file
